removing unused variables to reduce compiler warnings

+++ freex3d/src/lib/input/EAIEventsIn.c
@@ -358,7 +358,7 @@ char * fwl_EAI_handleBuffer(char *fromFront) {
 char * fwl_MIDI_handleBuffer(char *fromFront) {
 	/* memcp from fromFront to &EAI_BUFFER_CUR */
 	int eaiverbose;
-	int len = strlen(fromFront) ;
+	int len = (int) strlen(fromFront) ;
 	ttglobal tg = gglobal();
 	struct tEAIHelpers *th;
 	eaiverbose = tg->EAI_C_CommonFunctions.eaiverbose;
